Fix 6 HIGH/MEDIUM audit issues with 12 regression tests (Issues 696, 717, 729, 741, 742 + prior uncommitted fixes)
Batch 3 fixes:
- Issue 729: Empty password now rejected (was advisory-only warning)
- Issue 717: Hotkey slice panic on short strings — safe length check
- Issue 741: rao() as i64 overflow — upgraded to i128 in 4 locations
- Issue 696: Port overflow on cfg.port=65535 — checked_add + validation
- Issue 742: ip as u32 truncation — explicit mask before cast

Tests: 5 new ip_to_string tests, updated password validation tests,
proptest updated for Result return type. 421 lib tests pass.

+++ src/queries/cache.rs
@@ -199,7 +199,7 @@ pub fn diff(old: &Metagraph, new: &Metagraph) -> Vec<MetagraphDelta> {
                         },
                     });
                 }
-                let stake_diff = nn.stake.rao() as i64 - on.stake.rao() as i64;
+                let stake_diff = nn.stake.rao() as i128 - on.stake.rao() as i128;
                 if stake_diff.unsigned_abs() > 1_000_000_000 {
                     // > 1 TAO change
                     deltas.push(MetagraphDelta {
